Add a check for keystore files
The project needs to build without keystore files so that other users can build and test the app.
Currently the app fails to build if keystore related files are not present.
Add a check on keystore files and ignore signing if these files are not present.
The release builds workflow of CI/CD should create these files at run time using CI variables.